Print errors to stderr
If a path that does not exist is inputted the application crashed
with a null pointer panic.
As reported in https://github.com/walle/lll/pull/1.
This commit checks if the file info exists and if not outputs a user
friendly error message on stderr.
It does also check for any other error that could occur and outputs it
to stderr.
This commit also changes the signature of lll.ShouldSkip as the
error is not checked in the method.