Add tabwidth argument to lll
The tabwidth argument to lll makes it possible to configure how many
characters a tab should be counted as.  This is useful for reporting
long lines which are indented.

The default value is 1, meaning legacy behavior is kept.  By setting the
tabwidth, it's easy to get warnings about other tab width settings, such
as 4 and 8.