Add a new flag to control line-length of comment lines separately.
This allows users to have a longer line-length limit for code than for
natural-language text.

One example is a project that wants "most" code to fit in 80 columns,
but will allow the occassional long line.  The best way to implement
this in `lll` right now is to have a longer line limit (say, 100
columns) with an understanding the code-reviewers will make sure most
lines are not over 80 columns.  But we still want all comment lines to
fit within 80 columns, since there's no good reason for them not to.
This new flag allows for that.